Update path_optimizer.py
I managed to get all the random path generation to work, but unfortunately, it seems that this solution is unfeasible and will not work, due simply to how small of a chance there is of the astrobee properly navigating through the koz through sheer random movement. On the other hand, I have realized that most of the plotting and collision detection code can be repurposed in making a 3D A* algorithm, and that the only stuff I have to overhaul is the stuff I did yesterday which was the random path generation.