Debugging, Visualizations, & Outputting Data As Files
The program now saves all paths and other info that the A* algorithm outputs as Python pickled files. Additionally, I did a lot more debugging to sort out a few kinks (I basically reduced the problem down to the base case and built up from there), and I added some fancier visualizations to aid in the debugging. There's a lot more debugging and code profiling to do to optimize and fix this program, but for the most part, it's nearly functional.